chore: move TestSASsubmitTimeout to new test file
